API_cheese_image/requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.12.0 --no-cache-dir
pillow
numpy
PTL
flask-cors